feat: add schema for google sheets
