+++ frontend/src/Editor/DataSourceManager/SourceComponents/Googlesheets.schema.json
@@ -1,0 +1,15 @@
+{
+  "$schema": "https://json-schema.org/",
+  "$id": "https://tooljet.io/Airtable.schema.json",
+  "title": "Google Sheets datasource",
+  "description": "A schema defining google sheets datasource",
+  "type": "object",
+  "properties": {
+    "sheets": {
+      "$label": "",
+      "$key": "sheets",
+      "type": "react-component-google-sheets",
+      "description": "A component for google sheets"
+    }
+  }
+}
