+++ frontend/src/Editor/DataSourceManager/SourceComponents/index.js
@@ -1,10 +1,12 @@
-import React from 'react'
+import React from 'react';
 
 import DynamicForm from '@/_components/DynamicForm';
-import AirtableSchema from './Airtable.schema.json'
-import RestapiSchema from './Restapi.schema.json'
-import GraphqlSchema from './Graphql.schema.json'
-import StripeSchema from './Stripe.schema.json'
+import AirtableSchema from './Airtable.schema.json';
+import RestapiSchema from './Restapi.schema.json';
+import GraphqlSchema from './Graphql.schema.json';
+import StripeSchema from './Stripe.schema.json';
+import GooglesheetSchema from './Googlesheets.schema.json';
+
 import { Elasticsearch } from './Elasticsearch';
 import { Redis } from './Redis';
 import { Postgresql } from './Postgresql';
@@ -12,7 +14,7 @@ import { Mysql } from './Mysql';
 // import { Stripe } from './Stripe';
 import { Firestore } from './Firestore';
 // import { Restapi } from './Restapi';
-import { Googlesheets } from './Googlesheets';
+// import { Googlesheets } from './Googlesheets';
 import { Slack } from './Slack';
 import { Mongodb } from './Mongodb';
 import { Dynamodb } from './Dynamodb';
@@ -20,23 +22,24 @@ import { Dynamodb } from './Dynamodb';
 // import { Graphql } from './Graphql';
 import { Mssql } from './Mssql';
 
-const Airtable = ({...rest}) => <DynamicForm schema={AirtableSchema} {...rest} />
-const Restapi = ({...rest}) => <DynamicForm schema={RestapiSchema} {...rest} />
-const Graphql = ({...rest}) => <DynamicForm schema={GraphqlSchema} {...rest} />
-const Stripe = ({...rest}) => <DynamicForm schema={StripeSchema} {...rest} />
+const Airtable = ({ ...rest }) => <DynamicForm schema={AirtableSchema} {...rest} />;
+const Restapi = ({ ...rest }) => <DynamicForm schema={RestapiSchema} {...rest} />;
+const Graphql = ({ ...rest }) => <DynamicForm schema={GraphqlSchema} {...rest} />;
+const Stripe = ({ ...rest }) => <DynamicForm schema={StripeSchema} {...rest} />;
+const Googlesheets = ({ ...rest }) => <DynamicForm schema={GooglesheetSchema} {...rest} />;
 export const SourceComponents = {
-    Elasticsearch,
-    Redis,
-    Postgresql,
-    Mysql,
-    Stripe,
-    Firestore,
-    Restapi,
-    Googlesheets,
-    Slack,
-    Mongodb,
-    Dynamodb,
-    Airtable,
-    Graphql,
-    Mssql
+  Elasticsearch,
+  Redis,
+  Postgresql,
+  Mysql,
+  Stripe,
+  Firestore,
+  Restapi,
+  Googlesheets,
+  Slack,
+  Mongodb,
+  Dynamodb,
+  Airtable,
+  Graphql,
+  Mssql,
 };
